fixed issue with making caves / stairs
Fixed the issue described here:

"Using bricks on cave exit works fine, makes stairs up.
Using bricks on the cave entrance when on the surface, will create
stairs down.
Enter the stairs down, and the stairs up have reverted back to a cave
exit and the surface around it resets to what rpgwo would be /surface 0
Using bricks on the cave exit creates stairs up again... go up the
stairs and the stairs down has reverted back to cave entrance...
Seems we can have one or the other but not both. Quite the loop we got
going on here lol."